refactor: Rename keeper -> timetable
