--- escahp/src/main/java/com/esca/escahp/keeper/I_KeeperService.java
@@ -1,5 +1,0 @@
-package com.esca.escahp.keeper;
-
-public interface I_KeeperService {
-    
-}
