+++ escahp/src/main/java/com/esca/escahp/timetable/I_TimeTableService.java
@@ -1,0 +1,5 @@
+package com.esca.escahp.timetable;
+
+public interface I_TimeTableService {
+    
+}
