+++ escahp/src/main/java/com/esca/escahp/timetable/TimeTableService.java
@@ -1,16 +1,16 @@
-package com.esca.escahp.keeper;
+package com.esca.escahp.timetable;
 
 import com.esca.escahp.calendar.repository.CalendarRepository;
-import com.esca.escahp.keeper.repository.TimeTableRepositorty;
+import com.esca.escahp.timetable.repository.TimeTableRepositorty;
 
 import org.springframework.stereotype.Service;
 
 @Service
-public class KeeperService implements I_KeeperService{
+public class TimeTableService implements I_TimeTableService {
     private final CalendarRepository calendarRepository;
     private final TimeTableRepositorty timeTableRepositorty;
 
-    public KeeperService(CalendarRepository keeperRepository, TimeTableRepositorty timeTableRepositorty) {
+    public TimeTableService(CalendarRepository keeperRepository, TimeTableRepositorty timeTableRepositorty) {
         this.calendarRepository = keeperRepository;
         this.timeTableRepositorty = timeTableRepositorty;
     }
