+++ escahp/src/main/java/com/esca/escahp/timetable/dto/KeeperRequest.java
@@ -1,10 +1,10 @@
-package com.esca.escahp.keeper.dto;
+package com.esca.escahp.timetable.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class KeeperResponse {
+public class KeeperRequest {
 
 }
