+++ escahp/src/main/java/com/esca/escahp/timetable/entity/Times.java
@@ -1,4 +1,4 @@
-package com.esca.escahp.keeper.entity;
+package com.esca.escahp.timetable.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
