+++ escahp/src/main/java/com/esca/escahp/timetable/repository/TimeTableRepositorty.java
@@ -1,9 +1,9 @@
-package com.esca.escahp.keeper.repository;
+package com.esca.escahp.timetable.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.esca.escahp.keeper.entity.TimeTable;
+import com.esca.escahp.timetable.entity.TimeTable;
 
 @Repository
 public interface TimeTableRepositorty extends JpaRepository<TimeTable, Long>{
